requirements.txt
python-nmap==0.7.1
pytest==7.4.0
pytest-cov==4.1.0 
pandas>=2.0.0
numpy==1.26.4
scikit-learn>=1.3.0
joblib>=1.3.0 
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
flask==3.0.0
flask-cors==4.0.0
